Extract TestSceneKeyBindingInputQueueChange
